% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_endecrypt.R
\name{Decrypt}
\alias{Decrypt}
\title{Decrypt variable}
\usage{
Decrypt(x, prv.key, encryption.type = "aes256")
}
\arguments{
\item{x}{The variable to be decrypted}

\item{prv.key}{key object. The private key object. Use the LoadKey funtion to load the private key and assign it}

\item{encryption.type}{Refer PKI.encrypt function. The default is "aes256"}
}
\value{
Encrypted variable
}
\description{
Decrypt variable
}
\examples{
\dontrun{
Decrypt(x = Encrypted_object, prv.key = prvkey, encryption.type = "aes256")
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
