% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_endecrypt.R
\name{DecryptDf}
\alias{DecryptDf}
\title{Data frame decryption}
\usage{
DecryptDf(x, prv.key, encryption.type = "aes256")
}
\arguments{
\item{x}{encrypted data.frame. The dataset to decrypt}

\item{prv.key}{Key Object.
The private key object.
Use the LoadKey function to load the private key and assign it}

\item{encryption.type}{Refer PKI.encrypt function. The default is "aes256"}
}
\value{
Decrypted \code{data.frame}
}
\description{
Data frame decryption
}
\examples{
\dontrun{
DecryptDf(x = airquality, prv.key = prvkey, encryption.type = "aes256")
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
