% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_endecrypt.R
\name{Encrypt}
\alias{Encrypt}
\title{Encrypt variable}
\usage{
Encrypt(x, pub.key, encryption.type = "aes256")
}
\arguments{
\item{x}{The variable to be encrypted}

\item{pub.key}{key object. The public key object. Use the LoadKey funtion to load the public key and assign it}

\item{encryption.type}{Refer PKI.encrypt function. The default is "aes256"}
}
\value{
Encrypted variable
}
\description{
Encrypt variable
}
\examples{
\dontrun{
Encrypt(x = "abc", pub.key = pubkey, encryption.type = "aes256")
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
