% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_endecrypt.R
\name{EncryptCol}
\alias{EncryptCol}
\title{Column encryption}
\usage{
EncryptCol(x, pub.key, encryption.type = "aes256")
}
\arguments{
\item{x}{List or Column. The column to encrypt}

\item{pub.key}{Key Object. The public key object. Use the LoadKey function to load the public key and assign it}

\item{encryption.type}{Refer PKI.encrypt function. The default is "aes256"}
}
\value{
Encrypted list or column
}
\description{
Column encryption
}
\examples{
\dontrun{
EncryptCol(x = airquality$Ozone, pub.key = pubkey, encryption.type = "aes256")
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
