% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_endecrypt.R
\name{EncryptDf}
\alias{EncryptDf}
\title{Data frame encryption}
\usage{
EncryptDf(x, pub.key, encryption.type = "aes256")
}
\arguments{
\item{x}{data.frame. The dataset to encrypt}

\item{pub.key}{Key Object. The public key object. Use the LoadKey function to load the public key and assign it}

\item{encryption.type}{Refer PKI.encrypt function. The default is "aes256"}
}
\value{
Encrypted \code{data.frame}
}
\description{
Data frame encryption
}
\examples{
\dontrun{
EncryptDf(x = airquality, pub.key = pubkey, encryption.type = "aes256")
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
