% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endecrypt.R
\docType{package}
\name{endecrypt}
\alias{endecrypt}
\alias{endecrypt-package}
\title{endecrypt: Encryption and decryption}
\description{
Encrypt and decrypt dataframes, Columns, cells and vectors.
}
\author{
Revanth Nemani <raynemani@gmail.com>

Aditya Vikram <adyviky9@gmail.com>
}
