% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{LoadKey}
\alias{LoadKey}
\title{loads the generated keys}
\usage{
LoadKey(key.path, Private)
}
\arguments{
\item{key.path}{Specify the path of the key in "".
The file extensions should end in ".pem" or ".der"}

\item{Private}{Boolean.
If the key is a private key use TRUE.
If the key is public key use FALSE.}
}
\value{
Loads public and private keys from specific paths
}
\description{
loads the generated keys
}
\examples{
\dontrun{
LoadKey("Encription/private.pem", Private)
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
