% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{SaveGenKey}
\alias{SaveGenKey}
\title{Saves generated keys}
\usage{
SaveGenKey(bits = 2048, private.key.path, public.key.path)
}
\arguments{
\item{bits}{Number. Usually 2048}

\item{private.key.path}{path to save the private key.
It should have a file extension ".pem" or ".der"}

\item{public.key.path}{path to save the public key.
It should have a file extension ".pem" or ".der"}
}
\value{
Saves public and private keys in specified paths
}
\description{
Saves generated keys
}
\examples{
\dontrun{
SaveGenKey(
  bits = 2048,
  private.key.path = "Encription/private.pem",
  public.key.path = "Encription/public.pem"
)
}

}
\author{
"Revanth Nemani <raynemani@gmail.com>"

"Aditya Vikram <adyviky9@gmail.com>"
}
